pyproject.toml
[build-system]
requires = ["setuptools>=70.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "icarus-plotting"
version = "0.1.0"
description = "simple trajectory plotting for ARES-Icarus"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.9.0",
    "numpy>=1.20.0",
    "pandas>=2.0.0",
    "rerun-sdk>=0.23.0",
]